# docs/architecture/openapi/components/gm_openapi.yaml
openapi: "3.0.3"
info:
  title: API Greenhouse Monitoring
  version: 1.3.0
  description: |
    REST и WebSocket-совместимые операции мониторинга теплиц: метаданные (организация → теплицы → камеры), live- и архивное видео,
    онлайн-телеметрия датчиков, поиск событий и выгрузка рядов для графиков (FR-02), GraphQL для аналитики.

    **Справочники и ERD:** полный CRUD операционных сущностей (`greenhouses`, `sensors`, `digital_controllers`, `video_cameras`) и операции по строкам топологии
    с `sdate`/`edate` согласованы с логической моделью `docs/architecture/diagram/data/cnt_gm_db/20-metadata-database-structure.md`.

    **Идентификаторы:** в URL и JSON используются UUID, согласованные с метаданными `CNT_GM_DB` (PostgreSQL).
    Код теплицы `code` на объекте и в MQTT (`greenhouse_code` в интеграции edge) сопоставляется с полем `code` в ресурсе теплицы;
    см. `docs/architecture/diagram/data/cnt_gm_db/20-metadata-database-structure.md`, `docs/architecture/diagram/integration/70-edge-devices.md`.

    **Аутентификация:** доступ по bearer-токену; типично **OIDC** access token (JWT) из `CNT_GM_Identity_DB` / веб-клиента (FR-03).
    Роли: сценарии поиска событий и архива по событиям — **инженер**; отказ при недостаточных правах — `403`.

    **Нефункциональные ориентиры:** задержка live видео и датчиков до **30 с** (NFR-01, BR-R05); ответ поиска событий до **10 с** (NFR-02, BR-R06).

...
servers:
  - url: http://localhost:5025
...
security:
  - bearerAuth: []
...
paths:
...
  /api/v1/sensor-types:
    get:
      operationId: listSensorTypes
      summary: Список типов датчиков
      tags: [MasterCatalog.SensorTypes]
      parameters:
        - $ref: "#/components/parameters/Page"
        - $ref: "#/components/parameters/PageSize"
      responses:
        "200":
          description: Страница типов датчиков
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PagedSensorTypes"
        "401": { $ref: "#/components/responses/Unauthorized" }
        "403": { $ref: "#/components/responses/Forbidden" }
        "500": { $ref: "#/components/responses/InternalServerError" }
    post:
      operationId: createSensorType
      summary: Создать тип датчика
      tags: [MasterCatalog.SensorTypes]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SensorTypeCreateRequest"
      responses:
        "201":
          description: Тип датчика создан
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SensorType"
        "400": { $ref: "#/components/responses/BadRequest" }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "403": { $ref: "#/components/responses/Forbidden" }
        "409": { $ref: "#/components/responses/Conflict" }
        "500": { $ref: "#/components/responses/InternalServerError" }
  /api/v1/sensor-types/{id}:
    get:
      operationId: getSensorTypeById
      summary: Получить тип датчика
      tags: [MasterCatalog.SensorTypes]
      parameters:
        - $ref: "#/components/parameters/SensorTypeId"
      responses:
        "200":
          description: Тип датчика
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/GetSensorTypeByIdResponse"
        "401": { $ref: "#/components/responses/Unauthorized" }
        "403": { $ref: "#/components/responses/Forbidden" }
        "404": { $ref: "#/components/responses/NotFound" }
        "500": { $ref: "#/components/responses/InternalServerError" }
    patch:
      operationId: patchSensorType
      summary: Обновить тип датчика
      tags: [MasterCatalog.SensorTypes]
      parameters:
        - $ref: "#/components/parameters/SensorTypeId"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SensorTypePatchRequest"
      responses:
        "200":
          description: Тип датчика обновлён
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SensorType"
        "400": { $ref: "#/components/responses/BadRequest" }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "403": { $ref: "#/components/responses/Forbidden" }
        "404": { $ref: "#/components/responses/NotFound" }
        "409": { $ref: "#/components/responses/Conflict" }
        "500": { $ref: "#/components/responses/InternalServerError" }
    delete:
      operationId: deleteSensorType
      summary: Деактивировать тип датчика
      tags: [MasterCatalog.SensorTypes]
      parameters:
        - $ref: "#/components/parameters/SensorTypeId"
      responses:
        "204":
          description: Тип датчика деактивирован
        "401": { $ref: "#/components/responses/Unauthorized" }
        "403": { $ref: "#/components/responses/Forbidden" }
        "404": { $ref: "#/components/responses/NotFound" }
        "500": { $ref: "#/components/responses/InternalServerError" }
...
  /api/v1/sensors/{id}:
    get:
      operationId: getSensorById
      summary: Получить датчик
      tags: [Catalog.Sensors]
      parameters:
        - $ref: "#/components/parameters/SensorId"
      responses:
        "200":
          description: Датчик
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Sensor"
        "401": { $ref: "#/components/responses/Unauthorized" }
        "403": { $ref: "#/components/responses/Forbidden" }
        "404": { $ref: "#/components/responses/NotFound" }
        "500": { $ref: "#/components/responses/InternalServerError" }
    patch:
      operationId: patchSensor
      summary: Обновить датчик
      tags: [Catalog.Sensors]
      parameters:
        - $ref: "#/components/parameters/SensorId"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SensorPatchRequest"
      responses:
        "200":
          description: Датчик обновлён
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Sensor"
        "400": { $ref: "#/components/responses/BadRequest" }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "403": { $ref: "#/components/responses/Forbidden" }
        "404": { $ref: "#/components/responses/NotFound" }
        "500": { $ref: "#/components/responses/InternalServerError" }
    delete:
      operationId: deleteSensor
      summary: Деактивировать датчик
      tags: [Catalog.Sensors]
      parameters:
        - $ref: "#/components/parameters/SensorId"
      responses:
        "204":
          description: Датчик деактивирован
        "401": { $ref: "#/components/responses/Unauthorized" }
        "403": { $ref: "#/components/responses/Forbidden" }
        "404": { $ref: "#/components/responses/NotFound" }
        "500": { $ref: "#/components/responses/InternalServerError" }
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: |
        Access token (как правило JWT), выданный OIDC-провайдером после входа в веб-приложение (FR-03).
        Содержит `sub` пользователя; членство в организации и роли проверяются в `CNT_GM_WebAPI`.
  parameters:
...
    SensorTypeId:
      name: id
      in: path
      required: true
      description: Идентификатор типа датчика (`sensor_types.id`)
      schema:
        type: string
        format: uuid
...
    SensorId:
      name: id
      in: path
      required: true
      description: Идентификатор экземпляра датчика (`sensors.id`)
      schema:
        type: string
        format: uuid
...
    Page:
      name: page
      in: query
      required: false
      description: Номер страницы (с 1)
      schema:
        type: integer
        minimum: 1
        default: 1
    PageSize:
      name: page_size
      in: query
      required: false
      description: Размер страницы (ограничение сверху — по политике API, рекомендуется не более 100)
      schema:
        type: integer
        minimum: 1
        maximum: 100
        default: 25
  schemas:
...
    GetSensorTypeByIdResponse:
      type: object
      required: [data]
      description: Обёртка ответа GET типа датчика по id (полезная нагрузка в data)
      properties:
        data:
          $ref: "#/components/schemas/SensorType"
    SensorType:
      type: object
      required: [id, code, name, isActive]
      properties:
        id: { type: string, format: uuid }
        code: { type: string, maxLength: 64 }
        name: { type: string, maxLength: 256 }
        defaultUnit: { type: string, maxLength: 16, nullable: true }
        valueMin: { type: number, format: double, nullable: true }
        valueMax: { type: number, format: double, nullable: true }
        isActive: { type: boolean }
        createdAt: { type: string, format: date-time }
        updatedAt: { type: string, format: date-time }
    SensorTypeCreateRequest:
      type: object
      required: [code, name]
      properties:
        code: { type: string, maxLength: 64 }
        name: { type: string, maxLength: 256 }
        defaultUnit: { type: string, maxLength: 16, nullable: true }
        valueMin: { type: number, format: double, nullable: true }
        valueMax: { type: number, format: double, nullable: true }
    SensorTypePatchRequest:
      type: object
      properties:
        code: { type: string, maxLength: 64 }
        name: { type: string, maxLength: 256 }
        defaultUnit: { type: string, maxLength: 16, nullable: true }
        valueMin: { type: number, format: double, nullable: true }
        valueMax: { type: number, format: double, nullable: true }
        isActive: { type: boolean }
    PagedSensorTypes:
      type: object
      required: [items, page, pageSize, totalItems]
      properties:
        items:
          type: array
          items: { $ref: "#/components/schemas/SensorType" }
        page: { type: integer, minimum: 1 }
        pageSize: { type: integer, minimum: 1 }
        totalItems: { type: integer, minimum: 0 }
...
    Sensor:
      type: object
      required: [id, sensorTypeId, displayName, isActive, createdAt, updatedAt]
      properties:
        id: { type: string, format: uuid }
        sensorTypeId: { type: string, format: uuid, description: FK → `sensor_types.id` }
        displayName: { type: string, maxLength: 256 }
        installPosition: { type: string, maxLength: 128, nullable: true }
        isActive: { type: boolean }
        createdAt: { type: string, format: date-time }
        updatedAt: { type: string, format: date-time }
...
    SensorPatchRequest:
      type: object
      properties:
        sensorTypeId: { type: string, format: uuid }
        displayName: { type: string, maxLength: 256 }
        installPosition: { type: string, maxLength: 128, nullable: true }
        isActive: { type: boolean }
...
    ProblemDetails:
      type: object
      description: Упрощённое описание ошибки (можно заменить на RFC 7807 при внедрении на сервере)
      properties:
        title:
          type: string
        detail:
          type: string
        status:
          type: integer
  responses:
    BadRequest:
      description: Некорректный запрос (валидация, синтаксис выражения и т.д.)
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ProblemDetails"
    Unauthorized:
      description: Требуется аутентификация
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ProblemDetails"
    Forbidden:
      description: Доступ запрещён (недостаточно прав по роли или организации, FR-03)
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ProblemDetails"
    NotFound:
      description: Ресурс не найден
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ProblemDetails"
    InternalServerError:
      description: Внутренняя ошибка сервера
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ProblemDetails"
    Conflict:
      description: Конфликт данных (уникальность, пересечение периодов и т.п.)
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ProblemDetails"